Add useSubscription hook
This should be sufficient for the state management and syncrhonization
requirements of this application. A container registers a reducer which
receives the results of mutation and subscriptions (in case something is
changed in another open instance of the application). This reducer
handles merging the new "event" into state (such as an added task,
changed tasks, etc).